Enformer/requirements.txt
dm-sonnet==2.0.0
kipoiseq==0.5.2
numpy==1.19.5
pandas==1.2.3
tensorflow==2.5.0
tensorflow-hub==0.11.0
biopython==1.83